refactor test_etl and break out avro tests to test_avro folder
